+++ systemSimulation/entities/raspi/__init__.py
@@ -1,7 +1,6 @@
 from runtime.types import POWER_BOOTING, POWER_FAULT, POWER_OFF, POWER_READY
 
 from .control_program import NoopControlProgram
-from .delay_pipeline import DelayPipeline
 from .client import RaspiClient
 from .model import RaspiDelayModel
 from .entity import RaspiEntity, RaspiState